Moved the files into wisent subdir and check into tests subdir. Removed the automake shit from root. Created a python package.
