--- configure.ac
@@ -1,11 +1,0 @@
-dnl Process this file with autoconf to produce a configure script.
-dnl Copyright 2008, 2009  Jochen Voss
-
-AC_INIT(wisent, 0.6.2, voss@seehuhn.de, [], [http://seehuhn.de/pages/wisent])
-AM_INIT_AUTOMAKE
-
-dnl Python support
-AM_PATH_PYTHON
-
-AC_CONFIG_FILES([Makefile doc/Makefile examples/Makefile examples/css/Makefile])
-AC_OUTPUT
