+++ setup.cfg
@@ -1,0 +1,29 @@
+[metadata]
+name = wisent
+author = Jochen Voss
+url = https://github.com/seehuhn/wisent
+description = A package to parse and serizlize APT source.list files
+long_description = file: README
+keywords = parser generator
+license = GPL-2.0
+classifiers =
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Development Status :: 4 - Beta
+    Environment :: Other Environment
+    Intended Audience :: Developers
+    License :: OSI Approved
+    License :: OSI Approved :: GPL-2.0
+    Operating System :: OS Independent
+    Topic :: Software Development :: Libraries :: Python Modules
+
+[options]
+python_requires = >=3.4
+zip_safe = True
+include_package_data = True
+packages = wisent
+setup_requires = setuptools_scm;
+
+[options.entry_points]
+console_scripts =
+    wisent = wisent.__main__:main
